+++ OUDS/Core/Components/Sources/Actions/Button/OUDSButton.swift
@@ -70,7 +70,7 @@ import SwiftUI
 /// For standard or business-oriented journeys, keep the default corners.
 /// This evolution addresses the need for flexibility in adapting the design to some brand contexts.
 ///
-/// To activate the rounded button behavior, set to true the  `hasRoundedButtons` values of the `Tuning` object
+/// To activate the rounded button behavior, set to true the `hasRoundedButtons` values of the `Tuning` object
 /// in your theme configuration (if the theme exposes this property at init).
 /// Some themes do not have this flexibility like `SoshTheme` and `WireframeTheme`.
 ///