+++ OUDS/Core/Components/Sources/Controls/TextInput/Internal/LeadingIconContainer.swift
@@ -2,13 +2,13 @@
 // Software Name: OUDS iOS
 // SPDX-FileCopyrightText: Copyright (c) Orange SA
 // SPDX-License-Identifier: MIT
-// 
+//
 // This software is distributed under the MIT license,
 // the text of which is available at https://opensource.org/license/MIT/
 // or see the "LICENSE" file for more details.
-// 
+//
 // Authors: See CONTRIBUTORS.txt
-// Software description: A SwiftUI components library with code examples for Orange Unified Design System 
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
 //
 
 import OUDSTokensSemantic
@@ -19,7 +19,9 @@ struct LeadingIconContainer: View {
     // MARK: - Properties
 
     let leadingIcon: Image?
+    let flip: Bool
     let status: OUDSTextInput.Status
+
     @Environment(\.theme) private var theme
 
     // MARK: - Body
@@ -31,6 +33,7 @@ struct LeadingIconContainer: View {
             .aspectRatio(contentMode: .fit)
             .frame(height: theme.textInput.textInputSizeLeadingIcon, alignment: .center)
             .oudsForegroundColor(color)
+            .toFlip(flip)
     }
 
     // MARK: - Helper
@@ -38,9 +41,9 @@ struct LeadingIconContainer: View {
     private var color: MultipleColorSemanticTokens {
         switch status {
         case .default, .error, .loading, .readOnly:
-            return theme.colors.colorContentMuted
+            theme.colors.colorContentMuted
         case .disabled:
-            return theme.colors.colorActionDisabled
+            theme.colors.colorActionDisabled
         }
     }
 }