+++ OUDS/Core/Components/Sources/Controls/TextInput/Internal/TextInputBackgroundModifier.swift
@@ -2,13 +2,13 @@
 // Software Name: OUDS iOS
 // SPDX-FileCopyrightText: Copyright (c) Orange SA
 // SPDX-License-Identifier: MIT
-// 
+//
 // This software is distributed under the MIT license,
 // the text of which is available at https://opensource.org/license/MIT/
 // or see the "LICENSE" file for more details.
-// 
+//
 // Authors: See CONTRIBUTORS.txt
-// Software description: A SwiftUI components library with code examples for Orange Unified Design System 
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
 //
 
 import OUDSTokensSemantic
@@ -16,11 +16,12 @@ import SwiftUI
 
 struct TextInputBackgroundModifier: ViewModifier {
 
-    // MARK: - Stored properties
+    // MARK: - Properties
 
     let status: OUDSTextInput.Status
     let isOutlined: Bool
     let interactionState: TextInputInteractionState
+
     @Environment(\.theme) private var theme
 
     // MARK: - Body
@@ -44,29 +45,28 @@ struct TextInputBackgroundModifier: ViewModifier {
         case .default:
             switch interactionState {
             case .idle:
-                return theme.colors.colorActionSupportEnabled
+                theme.colors.colorActionSupportEnabled
             case .focused:
-                return theme.colors.colorActionSupportPressed
+                theme.colors.colorActionSupportPressed
             case .hover:
-                return theme.colors.colorActionSupportHover
+                theme.colors.colorActionSupportHover
             }
         case .error:
-            return theme.colors.colorSurfaceStatusNegativeMuted
+            theme.colors.colorSurfaceStatusNegativeMuted
         case .loading:
-            return theme.colors.colorActionSupportLoading
+            theme.colors.colorActionSupportLoading
         case .readOnly:
-            return nil
+            nil
         case .disabled:
-            return theme.colors.colorActionSupportDisabled
+            theme.colors.colorActionSupportDisabled
         }
     }
 
     private var outlinedColor: MultipleColorSemanticTokens? {
         if status == .readOnly {
-            return theme.colors.colorActionSupportDisabled
+            theme.colors.colorActionSupportDisabled
         } else {
-            return nil
+            nil
         }
     }
 }
-