+++ OUDS/Core/Components/Sources/Controls/TextInput/Internal/TextInputContainer.swift
@@ -2,13 +2,13 @@
 // Software Name: OUDS iOS
 // SPDX-FileCopyrightText: Copyright (c) Orange SA
 // SPDX-License-Identifier: MIT
-// 
+//
 // This software is distributed under the MIT license,
 // the text of which is available at https://opensource.org/license/MIT/
 // or see the "LICENSE" file for more details.
-// 
+//
 // Authors: See CONTRIBUTORS.txt
-// Software description: A SwiftUI components library with code examples for Orange Unified Design System 
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
 //
 
 import OUDSTokensSemantic
@@ -22,13 +22,15 @@ struct TextInputContainer: View {
     let text: Binding<String>
     let placeholder: OUDSTextInput.Placeholder?
     let leadingIcon: Image?
+    let flipIcon: Bool
     let trailingAction: OUDSTextInput.TrailingAction?
     let isOutlined: Bool
     let status: OUDSTextInput.Status
+    @State var hover: Bool = false
+
+    @FocusState private var focused: Bool
 
     @Environment(\.theme) private var theme
-    @FocusState private var focused: Bool
-    @State var hover: Bool = false
 
     // MARK: - Body
 
@@ -36,7 +38,7 @@ struct TextInputContainer: View {
         HStack(alignment: .center, spacing: theme.textInput.textInputSpaceColumnGapDefault) {
 
             // Leading icon container
-            LeadingIconContainer(leadingIcon: leadingIcon, status: status)
+            LeadingIconContainer(leadingIcon: leadingIcon, flip: flipIcon, status: status)
 
             // Text container
             VStack(alignment: .leading, spacing: theme.textInput.textInputSpaceRowGapLabelInput) {
@@ -51,24 +53,26 @@ struct TextInputContainer: View {
                     // Prefix container
                     if let placeholder,
                        let prefix = placeholder.prefix,
-                       !prefix.isEmpty, !placeholder.text.isEmpty {
+                       !prefix.isEmpty, !placeholder.text.isEmpty
+                    {
                         Text(prefix)
                             .typeLabelDefaultLarge(theme)
                             .oudsForegroundColor(prefixSuffixColor)
                     }
 
                     // Input text container
-                    Inputtext(label: textfieldLabel,
+                    InputText(label: textfieldLabel,
                               text: text,
                               labelAsPlaceholder: textfieldLabel == label,
                               status: status,
                               interactionState: interactionState)
                         .focused($focused, equals: true)
 
-                    // Sufix container
+                    // Suffix container
                     if let placeholder,
                        let suffix = placeholder.suffix,
-                       !suffix.isEmpty, !placeholder.text.isEmpty {
+                       !suffix.isEmpty, !placeholder.text.isEmpty
+                    {
 
                         Text(suffix)
                             .typeLabelDefaultLarge(theme)
@@ -82,11 +86,11 @@ struct TextInputContainer: View {
         }
         .padding(.vertical, theme.textInput.textInputSpacePaddingBlockDefault)
         .padding(.leading, theme.textInput.textInputSpacePaddingInlineDefault)
-        .padding(.trailing, trailing)
+        .padding(.trailing, trailingPadding)
         .frame(minHeight: theme.textInput.textInputSizeMinHeight, alignment: .leading)
         .modifier(TextInputBackgroundModifier(status: status, isOutlined: isOutlined, interactionState: interactionState))
-        .modifier(TextInputBoderModifier(status: status, isOutlined: isOutlined, interactionState: interactionState))
-        .onHover{ self.hover = $0 }
+        .modifier(TextInputBorderModifier(status: status, isOutlined: isOutlined, interactionState: interactionState))
+        .onHover { hover = $0 }
     }
 
     // MARK: - Helpers
@@ -106,11 +110,11 @@ struct TextInputContainer: View {
         return placeholder.text
     }
 
-    private var trailing: CGFloat {
+    private var trailingPadding: CGFloat {
         if trailingAction != nil || status == .error || status == .loading {
-            return theme.textInput.textInputSpacePaddingInlineTrailingAction
+            theme.textInput.textInputSpacePaddingInlineTrailingAction
         } else {
-            return theme.textInput.textInputSpacePaddingInlineDefault
+            theme.textInput.textInputSpacePaddingInlineDefault
         }
     }