+++ OUDS/Core/Components/Sources/Indicators/Badge/OUDSBadge.swift
@@ -49,7 +49,7 @@ import SwiftUI
 ///
 /// ## Design documentation
 ///
-/// [unified-design-system.orange.com](https://unified-design-system.orange.com/472794e18/p/73c701-components)
+/// [unified-design-system.orange.com](https://unified-design-system.orange.com/472794e18/p/698ea8-badge)
 ///
 /// ## Themes rendering
 ///
@@ -75,7 +75,7 @@ import SwiftUI
 ///
 /// - Version:1.1.0 (Figma component design version)
 /// - Since: 0.17.0
-public struct OUDSBadge: View { // TODO: #514 - Add hyperlink for badge documentation in documentation above
+public struct OUDSBadge: View {
 
     static let maxCount = 99